backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv